:bug: (src) pass e2e testing
