+++ README.md
@@ -7,6 +7,8 @@ Let's contribute OSS!!
 
 [https://expo.io/@nd-02110114/goofi-mobile](https://expo.io/@nd-02110114/goofi-mobile)
 
+<img  src="https://raw.githubusercontent.com/nd-02110114/goofi-mobile/master/gif/demo.gif" width="220" height="450"/>
+
 ## Development
 
 ### Prepare for the Detox
