+++ app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
     "name": "Goofi Mobile",
-    "version": "1.0.1",
+    "version": "2.0.0",
     "privacy": "public",
     "sdkVersion": "33.0.0",
     "orientation": "portrait",
